feat: Basic share implementation with materialui
